+++ src/inputs/wysiwyg/wysiwyg.interfaces.ts
@@ -1,0 +1,6 @@
+import {BaseInputFieldDataInterface} from '../base/baseInput.interfaces'
+
+export interface WysiwygFieldDataInterface extends BaseInputFieldDataInterface {
+	maxRows?: number
+	minRows?: number
+}